refactor http api into sub-folders
