+++ src/http/account/index.js
@@ -1,0 +1,15 @@
+module.exports = {
+    init: (app) => {
+        // get account info
+        app.get('/account/:name', (req, res) => {
+            if (!req.params.name) {
+                res.sendStatus(500)
+                return
+            }
+            db.collection('accounts').findOne({ name: req.params.name }, function (err, account) {
+                if (!account) res.sendStatus(404)
+                else res.send(account)
+            })
+        })
+    }
+}
