+++ src/http/blog/index.js
@@ -1,0 +1,34 @@
+module.exports = {
+    init: (app) => {
+        // get blog of user
+        app.get('/blog/:username', (req, res) => {
+            var username = req.params.username
+            db.collection('contents').find({ pa: null, author: username }, { sort: { ts: -1 }, limit: 50 }).toArray(function (err, contents) {
+                res.send(contents)
+            })
+        })
+        app.get('/blog/:username/:author/:link', (req, res) => {
+            db.collection('contents').findOne({
+                $and: [
+                    { author: req.params.author },
+                    { link: req.params.link }
+                ]
+            }, function (err, content) {
+                if (err || !content) {
+                    res.send([])
+                    return
+                }
+                var username = req.params.username
+                db.collection('contents').find({
+                    $and: [
+                        { pa: null },
+                        { author: username },
+                        { ts: { $lte: content.ts } }
+                    ]
+                }, { sort: { ts: -1 }, limit: 50 }).toArray(function (err, contents) {
+                    res.send(contents)
+                })
+            })
+        })
+    }
+}
