+++ src/http/config/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    init: (app) => {
+        // get current chain config
+        app.get('/config', (req, res) => {
+            res.send(config)
+        })
+    }
+}
