+++ src/http/new/index.js
@@ -1,0 +1,31 @@
+module.exports = {
+    init: (app) => {
+        // get new contents
+        app.get('/new', (req, res) => {
+            db.collection('contents').find({ pa: null }, { sort: { ts: -1 }, limit: 50 }).toArray(function (err, contents) {
+                res.send(contents)
+            })
+        })
+        app.get('/new/:author/:link', (req, res) => {
+            db.collection('contents').findOne({
+                $and: [
+                    { author: req.params.author },
+                    { link: req.params.link }
+                ]
+            }, function (err, content) {
+                if (!content) {
+                    res.sendStatus(404)
+                    return
+                }
+                db.collection('contents').find({
+                    $and: [
+                        { pa: null },
+                        { ts: { $lte: content.ts } }
+                    ]
+                }, { sort: { ts: -1 }, limit: 50 }).toArray(function (err, contents) {
+                    res.send(contents)
+                })
+            })
+        })
+    }
+}
