+++ src/http/newKeyPair/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    init: (app) => {
+        // generate a new key pair
+        app.get('/newKeyPair', (req, res) => {
+            res.send(chain.getNewKeyPair())
+        })
+    }
+}
