+++ src/http/rewardPool/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+    init: (app) => {
+        // check econ data
+        app.get('/rewardPool', (req, res) => {
+            if (eco.lastRewardPool)
+                res.send(eco.lastRewardPool)
+            else res.send({})
+        })
+    }
+}
