+++ src/http/schedule/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    init: (app) => {
+        // look at the miner schedule
+        app.get('/schedule', (req, res) => {
+            res.send(chain.schedule)
+        })
+    }
+}
